requirements.txt
colorama==0.3.7
django-cors-headers==2.0.0
django-rest-swagger==2.1.1
django-sslserver==0.19
Django==1.10.2
djangorestframework-jwt==1.9.0
djangorestframework==3.5.1
freezegun==0.3.8
httpie==0.9.6
humanfriendly==2.4
py==1.4.31
Pygments==2.1.3
pytest-cov==2.4.0
pytest-django==3.1.2
pytest==3.0.6
requests==2.11.1
tox==2.6.0
wheel==0.24.0